Update the README
We were going to make this project a complete tracking system for the
Replicated vendor portal. It turned out this was not feasible in a
single project.

Instead, lets build a command line interface for interacting with the
API that we can build off of.

Update the readme (and repo metadata) to reflect the new goal.